+++ nilearn/image/image.py
@@ -775,18 +775,81 @@ def new_img_like(ref_niimg, data, affine=None, copy_header=False):
     return klass(data, affine, header=header)
 
 
-def threshold_img(img, threshold, mask_img=None, copy=True):
+def _apply_cluster_size_threshold(arr, cluster_threshold, copy=True):
+    """Apply cluster-extent thresholding to an array that has already been
+    voxel-wise thresholded.
+
+    Parameters
+    ----------
+    arr : :obj:`numpy.ndarray` of shape (X, Y, Z)
+        3D array that has been thresholded at the voxel level.
+    cluster_threshold : :obj:`float`
+        Cluster-size threshold, in voxels, to apply to ``arr``.
+    copy : :obj:`bool`, optional
+        Whether to copy the array before modifying it or not.
+        Default is True.
+
+    Returns
+    -------
+    arr : :obj:`numpy.ndarray` of shape (X, Y, Z)
+        Cluster-extent thresholded array.
+
+    Notes
+    -----
+    Clusters are defined in a bi-sided manner;
+    both negative and positive clusters are evaluated,
+    but this is done separately for each sign.
+
+    Clusters are defined using 6-connectivity, also known as NN1 (in AFNI) or
+    "faces" connectivity.
+    """
+    assert arr.ndim == 3
+
+    if copy:
+        arr = arr.copy()
+
+    # Define array for 6-connectivity, aka NN1 or "faces"
+    conn_mat = np.zeros((3, 3, 3), int)
+    conn_mat[:, 1, 1] = 1
+    conn_mat[1, :, 1] = 1
+    conn_mat[1, 1, :] = 1
+
+    for sign in np.sign(arr):
+        # Binarize using one-sided cluster-defining threshold
+        binarized = ((arr * sign) > 0).astype(int)
+
+        # Apply cluster threshold
+        label_map = ndimage.measurements.label(binarized, conn_mat)[0]
+        clust_ids = sorted(list(np.unique(label_map)[1:]))
+        for c_val in clust_ids:
+            if np.sum(label_map == c_val) < cluster_threshold:
+                arr[label_map == c_val] = 0
+
+    return arr
+
+
+def threshold_img(
+    img,
+    threshold,
+    cluster_threshold=0,
+    two_sided=True,
+    mask_img=None,
+    copy=True,
+):
     """Threshold the given input image, mostly statistical or atlas images.
 
     Thresholding can be done based on direct image intensities or selection
     threshold with given percentile.
+
+    .. versionchanged:: 0.8.2
+        New ``cluster_threshold`` and ``two_sided`` parameters added.
 
     .. versionadded:: 0.2
 
     Parameters
     ----------
     img : a 3D/4D Niimg-like object
-        Image contains of statistical or atlas maps which should be thresholded.
+        Image containing statistical or atlas maps which should be thresholded.
 
     threshold : :obj:`float` or :obj:`str`
         If float, we threshold the image based on image intensities meaning
@@ -797,6 +860,20 @@ def threshold_img(img, threshold, mask_img=None, copy=True):
         based on the score obtained using this percentile on the image data. The
         voxels which have intensities greater than this score will be kept.
         The given string should be within the range of "0%" to "100%".
+
+    cluster_threshold : :obj:`float`, optional
+        Cluster size threshold, in voxels. In the returned thresholded map,
+        sets of connected voxels (``clusters``) with size smaller
+        than this number will be removed. Default=0.
+
+        .. versionadded:: 0.8.2
+
+    two_sided : :obj:`bool`, optional
+        Whether the thresholding should yield both positive and negative
+        part of the maps.
+        Default=True.
+
+        .. versionadded:: 0.8.2
 
     mask_img : Niimg-like object, default None, optional
         Mask image applied to mask the input data.
@@ -813,7 +890,9 @@ def threshold_img(img, threshold, mask_img=None, copy=True):
 
     See also
     --------
-    nilearn.glm.threshold_stats_img
+    nilearn.glm.threshold_stats_img :
+        Threshold a statistical image using the alpha value, optionally with
+        false positive control.
 
     """
     from . import resampling
@@ -834,18 +913,40 @@ def threshold_img(img, threshold, mask_img=None, copy=True):
         # Set as 0 for the values which are outside of the mask
         img_data[mask_data == 0.] = 0.
 
-    if threshold is None:
-        raise ValueError("The input parameter 'threshold' is empty. "
-                         "Please give either a float value or a string as e.g. '90%'.")
+    cutoff_threshold = check_threshold(
+        threshold,
+        img_data,
+        percentile_func=scoreatpercentile,
+        name='threshold',
+    )
+
+    # Apply threshold
+    if two_sided:
+        img_data[np.abs(img_data) < cutoff_threshold] = 0.
     else:
-        cutoff_threshold = check_threshold(threshold, img_data,
-                                           percentile_func=scoreatpercentile,
-                                           name='threshold')
+        img_data[img_data < cutoff_threshold] = 0.
 
-    img_data[np.abs(img_data) < cutoff_threshold] = 0.
-    threshold_img = new_img_like(img, img_data, affine)
+    # Expand to 4D to support both 3D and 4D
+    expand_to_4d = img_data.ndim == 3
+    if expand_to_4d:
+        img_data = img_data[:, :, :, None]
 
-    return threshold_img
+    # Perform cluster thresholding, if requested
+    if cluster_threshold > 0:
+        for i_vol in range(img_data.shape[3]):
+            img_data[..., i_vol] = _apply_cluster_size_threshold(
+                img_data[..., i_vol],
+                cluster_threshold,
+            )
+
+    if expand_to_4d:
+        # Reduce back to 3D
+        img_data = img_data[:, :, :, 0]
+
+    # Reconstitute img object
+    thresholded_img = new_img_like(img, img_data, affine)
+
+    return thresholded_img
 
 
 def math_img(formula, **imgs):